refactor: extract account page details into Profile component
